# Conflicts:
#	administrator/components/com_emundus/script.com_emundus.php
#	components/com_emundus/controllers/formbuilder.php
#	components/com_emundus/language/en-GB/en-GB.com_emundus.ini
#	components/com_emundus/language/fr-FR/fr-FR.com_emundus.ini
#	components/com_emundus/views/form/tmpl/formbuilder.php
#	media/com_emundus_vue/app_emundus.css
#	media/com_emundus_vue/app_emundus.js
#	media/com_emundus_vue/app_emundus.js.map
#	plugins/fabrik_element/panel/panel-min.js
#	plugins/fabrik_element/panel/panel.js
#	templates/g5_helium/custom/scss/main.compiled.css

+++ plugins/fabrik_cron/emunduslogsandmessagespurge/emunduslogsandmessagespurge.php
@@ -75,17 +75,29 @@ class PlgFabrik_Cronemunduslogsandmessagespurge extends PlgFabrik_Cron{
 
 		if ($export)
 		{
-			$zip_filename = JPATH_SITE . '/tmp/backup_logs_and_messages_' . date('Y-m-d_H-i-s') . '.zip';
-			$zip          = new ZipArchive();
-			if ($zip->open($zip_filename, ZipArchive::CREATE) === true)
+			$filename_logs     = $m_logs->exportLogsBeforeADate($now);
+			$filename_messages = $m_messages->exportMessagesBeforeADate($now);
+
+			if(!empty($filename_messages) || !empty($filename_logs))
 			{
-				$filename_logs     = $m_logs->exportLogsBeforeADate($now);
-				$filename_messages = $m_messages->exportMessagesBeforeADate($now);
-				$zip->addFile($filename_logs, basename($filename_logs));
-				$zip->addFile($filename_messages, basename($filename_messages));
-				$zip->close();
-				unlink($filename_messages);
-				unlink($filename_logs);
+				$zip_filename = JPATH_SITE . '/tmp/backup_logs_and_messages_' . date('Y-m-d_H-i-s') . '.zip';
+				$zip          = new ZipArchive();
+				if ($zip->open($zip_filename, ZipArchive::CREATE) === true)
+				{
+					if (!empty($filename_logs) && file_exists($filename_logs))
+					{
+						$zip->addFile($filename_logs, basename($filename_logs));
+						unlink($filename_logs);
+					}
+
+					if (!empty($filename_messages) && file_exists($filename_messages))
+					{
+						$zip->addFile($filename_messages, basename($filename_messages));
+						unlink($filename_messages);
+					}
+
+					$zip->close();
+				}
 			}
 		}
